+++ src/token.cpp
@@ -14,16 +14,42 @@ std::string ToString(TokenType type) {
             return "TOKEN_INICIO";
         case TokenType::FimAlgoritmo:
             return "TOKEN_FIMALGORITMO";
+        case TokenType::Se:
+            return "TOKEN_SE";
+        case TokenType::Entao:
+            return "TOKEN_ENTAO";
+        case TokenType::Senao:
+            return "TOKEN_SENAO";
+        case TokenType::FimSe:
+            return "TOKEN_FIMSE";
+        case TokenType::Enquanto:
+            return "TOKEN_ENQUANTO";
+        case TokenType::Faca:
+            return "TOKEN_FACA";
+        case TokenType::FimEnquanto:
+            return "TOKEN_FIMENQUANTO";
         case TokenType::Inteiro:
             return "TOKEN_INTEIRO";
         case TokenType::Real:
             return "TOKEN_REAL";
         case TokenType::Texto:
             return "TOKEN_TEXTO";
+        case TokenType::Logico:
+            return "TOKEN_LOGICO";
+        case TokenType::Verdadeiro:
+            return "TOKEN_VERDADEIRO";
+        case TokenType::Falso:
+            return "TOKEN_FALSO";
         case TokenType::Leia:
             return "TOKEN_LEIA";
         case TokenType::Escreva:
             return "TOKEN_ESCREVA";
+        case TokenType::E:
+            return "TOKEN_E";
+        case TokenType::Ou:
+            return "TOKEN_OU";
+        case TokenType::Nao:
+            return "TOKEN_NAO";
         case TokenType::Plus:
             return "TOKEN_PLUS";
         case TokenType::Minus:
@@ -44,6 +70,8 @@ std::string ToString(TokenType type) {
             return "TOKEN_LESS_EQUAL";
         case TokenType::GreaterEqual:
             return "TOKEN_GREATER_EQUAL";
+        case TokenType::NotEqual:
+            return "TOKEN_NOT_EQUAL";
         case TokenType::Colon:
             return "TOKEN_COLON";
         case TokenType::Comma: